requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-contact-form==1.0
django-endless-pagination==2.0
django-markdown==0.8.4
django-toolbelt==0.0.1
gunicorn==19.3.0
Markdown==2.6.1
psycopg2==2.6
static3==0.5.1
whitenoise==1.0.6
